[Refactor] Rename OLAP tablet sink files
