+++ be/src/exec/data_sinks/data_sink_factory.cpp
@@ -20,7 +20,7 @@
 #include "data_sink/external/blackhole_table_sink.h"
 #include "data_sink/external/export_sink.h"
 #include "data_sink/tablet/multi_olap_table_sink.h"
-#include "data_sink/tablet/tablet_sink.h"
+#include "data_sink/tablet/olap_table_sink.h"
 #include "exec/data_sinks/dictionary_cache_sink.h"
 #include "exec/data_sinks/hive_table_sink.h"
 #include "exec_primitive/data_sink.h"
